requirements.txt
asgiref==3.6.0
Django==4.1.5
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-registration-redux==2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
